Refactored all OCEs into separate folders/files in Testing
