+++ imports/api/Testing/spring18study/four_interaction_structures.js
@@ -1,0 +1,12 @@
+import SAMESITUATION_SAMETIME from "./samesituation_sametime";
+import SAMESITUATION_ACROSSTIME from "./samesituation_acrosstime";
+import SHARED_GOALS from "./shared_goals";
+import BUILD_UPON_OTHERS_CONTRIBUTIONS from "../summer18study/mechanismRichStory";
+
+export default FOUR_INTERACTION_STRUCTURES_EXPERIENCES = Object.assign(
+  {},
+  SAMESITUATION_SAMETIME,
+  SAMESITUATION_ACROSSTIME,
+  SHARED_GOALS,
+  BUILD_UPON_OTHERS_CONTRIBUTIONS
+);
