+++ imports/api/Testing/summer18study/summer18experiences.js
@@ -1,0 +1,11 @@
+import MECHANISMPOOR_SAMESITUATION from "./mechanismPoorSameSituation";
+import MECHANISMRICH_SAMESITUATION from "./mechanismRichSameSituation";
+import MECHANISMPOOR_STORY from "./mechanismPoorStory";
+import MECHANISMRICH_STORY from "./mechanismRichStory";
+
+export default SUMMER18_EXPERIENCES = Object.assign({},
+  MECHANISMPOOR_SAMESITUATION,
+  MECHANISMRICH_SAMESITUATION,
+  MECHANISMPOOR_STORY,
+  MECHANISMRICH_STORY 
+);
