+++ imports/api/Testing/testingconstants.js
@@ -1,2167 +1,24 @@
-import { Meteor } from "meteor/meteor";
-
-import { Submissions } from "../OCEManager/currentNeeds";
-
-import { addContribution } from '../OCEManager/OCEs/methods';
-import {notify, notifyUsersInIncident, notifyUsersInNeed} from "../OpportunisticCoordinator/server/noticationMethods";
-import {Incidents} from "../OCEManager/OCEs/experiences";
-import {getDetectorUniqueKey, addStaticAffordanceToNeeds} from "./oce_api_helpers";
-
 import {LOCATIONS} from "./LOCATIONS";
 import {USERS} from "./USERS";
 import {DETECTORS} from "./DETECTORS";
 
-/**
- * Create Storytime Helper.
- *
- * @param version [number] determines which detector comes first
- * @return {{_id: string, name: string, participateTemplate: string, resultsTemplate: string, contributionTypes: *[], description: string, notificationText: string, callbacks: *[]}}
- */
-function createStorytime(version) {
-  // setup places and detectors for storytime
-  let places = ["niceish_day", "beer", "train", "forest", "dinning_hall", "castle", "field", "gym"];
-  let detectorUniqueKeys = places.map((x) => { return Random.id(); });
-  let detectorNames = [];
-  let dropdownText = [
-    'Swirling Clouds',
-    'Drinking butterbeer',
-    'Hogwarts Express',
-    'Forbidden Forest',
-    'Dinner at the Great Hall',
-    'Hogwarts Castle',
-    'Quidditch Pitch',
-    'Training in the Room of Requirement ',
-  ];
-
-  _.forEach(places, (place, i) => {
-    let newVars = JSON.parse(JSON.stringify(DETECTORS[place]['variables']));
-    newVars.push(`var participatedInStorytime${version};`);
-    newVars.push(`var mechanismRich;`);
-    let newRules = JSON.parse(JSON.stringify(DETECTORS[place]['rules']));
-    // modify last detector rule
-    // when rules has a flat structure where rules.length == 1, last rule is the predicate
-    // i.e. ['(diners || restaurants || cafeteria || food_court);']
-    // when rules have a nested structure where rules.length > 1, last rule is the predicate
-    // i.e. ['worship_places = (buddhist_temples || churches);', '(worship_places || landmarks);']
-    let lastRule = newRules.pop();
-    // each rule has a `;` at end, i.e. (rain && park);
-    // in order to modify the rule, must add predicate preceding the rule
-    let lastRuleNoSemicolon = lastRule.split(';')[0];
-    lastRule = `(mechanismRich && (!participatedInStorytime${version} && (${lastRuleNoSemicolon})));`;
-    newRules.push(lastRule);
-
-    let detectorName = `${place}_storytime${version}_mechanismRich`;
-    detectorNames.push(detectorName);
-
-    DETECTORS[detectorName] = {
-      '_id': detectorUniqueKeys[i],
-      'description': `${DETECTORS[place].description} storytime${version} mechanismRich`,
-      'variables': newVars,
-      'rules': newRules
-    };
-  });
-
-  // Don't assume the Random detectorUniqueKeys we created actually exist
-  detectorUniqueKeys = detectorNames.map((name) => { return getDetectorUniqueKey(DETECTORS[name]); });
-  let DROPDOWN_OPTIONS = _.zip(dropdownText, detectorUniqueKeys);
-  // create story starting point
-  let sentences = [
-    'Ron looked up at the clouds swirling above him.',
-    'Hermoine looked into her goblet, hardly realizing the unusual color of the concoction she was being forced to drink.',
-    'Harry prepared himself for a lunge, and then dove forward towards the Platform 9 3/4 wall.',
-    'The wizard looked down at their feet, hardly believing the magical plants growing in the Forbidden Forest.',
-    'Any young wizard who has their first meal in the Hogwarts Great Hall has to be surprised by the type of food on the menu.',
-    'Hogwarts castle had looked so good in photos, but this new wizard looked up at it unimpressed.',
-    'Harry Potter saw the snitch diving towards the ground. He aimed his broom towards the grassy ground and followed, reaching his hand out to grab it.',
-    'The new wizard of Dumbledore\'s Army was training very hard in the Room of Requirement.'
-  ];
-  let firstSentence = sentences[version];
-  let [firstSituation, firstDetector] = DROPDOWN_OPTIONS[version];
-  // notify users when story is complete
-  let sendNotification = function (sub) {
-    let uids = Submissions.find({iid: sub.iid}).fetch().map(function (x) {
-      return x.uid;
-    });
-
-    notify(uids, sub.iid, 'Our story is finally complete. Click here to read it!',
-      '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-  };
-
-  /**
-   * NOTE: if callback depends on any variables defined outside of its scope, we must use some solution so that
-   * the variables values are substituted into the callback.toString()
-   *
-   * For a dynamic code generation solution,
-   * @see https://stackoverflow.com/questions/29182244/convert-a-string-to-a-template-string
-   * @see https://medium.com/@oprearocks/serializing-object-methods-using-es6-template-strings-and-eval-c77c894651f0
-   * @param sub
-   */
-  let storytimeCallback = function (sub) {
-    Meteor.users.update({
-      _id: sub.uid
-    }, {
-      $set: {
-        ['profile.staticAffordances.participatedInStorytime${version}']: true
-      }
-    });
-
-    // set affordances for storytime
-    let affordance = sub.content.affordance;
-
-    // HACKY TEMPLATE DYNAMIC CODE GENERATION
-    let options = eval('${JSON.stringify(DROPDOWN_OPTIONS)}');
-
-    let [situation, detectorUniqueKey] = options.find(function(x) {
-      return x[1] === affordance;
-    });
-
-    // options = options.filter(function (x) {
-    //   return x[1] !== affordance;
-    // });
-
-    // add need if not all pages are done
-    let needName = 'page' + Random.id(3);
-    if (cb.numberOfSubmissions() === 7) {
-      needName = 'pageFinal'
-    }
-
-    // create and add contribution
-    let contribution = {
-      needName: needName,
-      situation: {
-        detector: affordance,
-        number: '1'
-      },
-      toPass: {
-        instruction: sub.content.sentence,
-        situation: situation,
-        previousUserId: sub.uid,
-        dropdownChoices: {
-          name: 'affordance',
-          options: options
-        }
-      },
-      numberNeeded: 1,
-      notificationDelay: 90
-    };
-
-    addContribution(sub.iid, contribution);
-  };
-
-  // FIXME(rlouie): Can't have more than version 0,1,2
-  let exp_names = [
-    "A Ron Weasley Story",
-    "A Hermoine Granger Story",
-    "A Harry Potter Story"
-  ];
-
-  // create and return storytime experience
-  return {
-    _id: Random.id(),
-    name: exp_names[version],
-    participateTemplate: 'storyPage',
-    resultsTemplate: 'storybook',
-    contributionTypes: [{
-      needName: 'pageOne',
-      situation: {
-        detector: firstDetector,
-        number: '1'
-      },
-      toPass: {
-        instruction: firstSentence,
-        firstSentence: firstSentence,
-        situation: firstSituation,
-        dropdownChoices: {
-          name: 'affordance',
-          options: DROPDOWN_OPTIONS
-        }
-      },
-      numberNeeded: 1,
-      notificationDelay: 90
-    }],
-    description: 'We\'re writing a Harry Potter spin-off story',
-    notificationText: 'View this and other available experiences',
-    callbacks: [
-      {
-        trigger: 'cb.newSubmission() && (cb.numberOfSubmissions() <= 7)',
-        // substitute any variables used outside of the callback function scope
-        function: eval('`' + storytimeCallback.toString() + '`'),
-      },
-      {
-        trigger: 'cb.incidentFinished()',
-        function: sendNotification.toString()
-      }]
-  };
-}
-
-
-
-const createIndependentStorybook = () => {
-
-  let place_situation_delay = [
-    ["niceish_day",'Swirling Clouds', 5],
-    ["beer", 'Drinking butterbeer', 90],
-    ["train", 'Hogwarts Express', 90],
-    ["forest",'Forbidden Forest', 90],
-    ["dinning_hall",'Dinner at the Great Hall', 90],
-    ["castle",'Hogwarts Castle', 90],
-    ["field",'Quidditch Pitch', 90],
-    ["gym",'Training in the Room of Requirement', 90]
-  ];
-
-  return {
-    _id: Random.id(),
-    name: 'Humans of Hogwarts',
-    participateTemplate: 'storyPage_noInterdependence',
-    resultsTemplate: 'storyBook_noInterdependence',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', (function(place_situation_delay) {
-      return place_situation_delay.map((x) => {
-        let [place, situation, delay] = x;
-        return {
-          needName: situation,
-          situation: {
-            detector: getDetectorUniqueKey(DETECTORS[place]),
-            number: '1',
-          },
-          toPass: {
-            situation: situation
-          },
-          numberNeeded: 2,
-          notificationDelay: delay
-        }
-      });
-    })(place_situation_delay)),
-    description: 'We\'re writing a Harry Potter spin-off story',
-    notificationText: 'View this and other available experiences',
-    callbacks: [
-      {
-        trigger: 'cb.newSubmission()',
-        function: (notifyUsersInIncident('Someone added to Humans of Hogwarts',
-          'View photos and lines others have created')).toString()
-      },
-      {
-        trigger: 'cb.incidentFinished()',
-        function: (notifyUsersInIncident('Humans of Hogwarts has finished',
-          "View everyone's photos and lines that were contributed")).toString()
-      }]
-  };
-};
-
-function createBumped() {
-  let experience = {
-    name: 'Bumped',
-    participateTemplate: 'bumped',
-    resultsTemplate: 'bumpedResults',
-    contributionTypes: [],
-    description: 'You just virtually bumped into someone!',
-    notificationText: 'You just virtually bumped into someone!',
-    callbacks: []
-  };
-
-  let bumpedCallback = function (sub) {
-    console.log("calling the bumped callback!!!");
-
-    let otherSub = Submissions.findOne({
-      uid: {
-        $ne: sub.uid
-      },
-      iid: sub.iid,
-      needName: sub.needName
-    });
-
-    notify([sub.uid, otherSub.uid], sub.iid, 'See a photo from who you virtually bumped into!', '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-  };
-
-  let relationships = ['lovesDTR', 'lovesGarrett', 'lovesMeg', 'lovesMaxine'];
-  let places = [
-    ["bar", "at a bar"],  // like Cheers!
-    ["coffee", "at a coffee shop"],
-    ["grocery", "at a grocery store"],
-    ["restaurant", "at a restaurant"],
-    ["train", "commuting"],
-    ["exercising", "exercising"]
-  ];
-  _.forEach(relationships, (relationship) => {
-    _.forEach(places, (place) => {
-
-      let newVars = JSON.parse(JSON.stringify(DETECTORS[place[0]]['variables']));
-      newVars.push('var ' + relationship + ';');
-
-      let newRules = JSON.parse(JSON.stringify(DETECTORS[place[0]]['rules']));
-      // modify last detector rule
-      // when rules has a flat structure where rules.length == 1, last rule is the predicate
-      // i.e. ['(diners || restaurants || cafeteria || food_court);']
-      // when rules have a nested structure where rules.length > 1, last rule is the predicate
-      // i.e. ['worship_places = (buddhist_temples || churches);', '(worship_places || landmarks);']
-      let lastRule = newRules.pop();
-      // each rule has a `;` at end, i.e. (rain && park);
-      // in order to modify the rule, must add relationship predicate preceding the rule
-      let lastRuleNoSemicolon = lastRule.split(';')[0];
-      lastRule = `(${relationship} && (${lastRuleNoSemicolon}));`;
-      newRules.push(lastRule);
-
-      let detector = {
-        '_id': Random.id(),
-        'description': DETECTORS[place[0]].description + relationship,
-        'variables': newVars,
-        'rules': newRules
-      };
-      DETECTORS[place[0] + relationship] = detector;
-
-      for (let i = 0; i < 1; i++) {
-        let need = {
-          needName: place[0] + relationship + i,
-          situation: {
-            detector: getDetectorUniqueKey(detector),
-            number: '2'
-          },
-          toPass: {
-            instruction: 'You are at a  ' + place[1] + ' at the same time as '
-          },
-          numberNeeded: 2,
-          notificationDelay: 90
-        };
-
-        let callback = {
-          trigger: 'cb.numberOfSubmissions(\'' + place[0] + relationship + i + '\') === 2',
-          function: bumpedCallback.toString(),
-        };
-
-        experience.contributionTypes.push(need);
-        experience.callbacks.push(callback)
-      }
-    })
-  });
-
-  return experience;
-}
-
-/** createHalfHalf
- *
- * @param numberInSituation [Integer] number of people that need to be in the same situation at the same time
- * @param notificationDelay [Integer] notificationDelay for all places
- * @returns {{name: string, participateTemplate: string, resultsTemplate: string, contributionTypes: Array, description: string, notificationText: string, callbacks: Array}}
- */
-const createHalfHalf = function(
-  {
-    numberInSituation = 1,
-    notificationDelay = 90
-  } = {}
-) {
-  let experience = {
-    name: 'Half Half Bumped',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: [],
-    description: 'Participate in HalfHalf Travel: Capture your side of the story',
-    notificationText: 'Participate in HalfHalf Travel: Capture your side of the story',
-    callbacks: []
-  };
-
-
-  let completedCallback = function(sub) {
-    console.log("Another pair of halves completed a photo");
-
-    let submissions = Submissions.find({
-      iid: sub.iid,
-      needName: sub.needName
-    }).fetch();
-
-    let participants = submissions.map((submission) => { return submission.uid; });
-
-    notify(participants, sub.iid,
-      `Two people completed a half half photo`,
-      `See the results under ${sub.needName}`,
-      '/apicustomresults/' + sub.iid + '/' + sub.eid);
-  };
-
-  let places = [
-    ["bar", "at a bar", notificationDelay],
-    ["coffee", "at a coffee shop", notificationDelay],
-    ["grocery", "at a grocery store", notificationDelay],
-    ["restaurant", "at a restaurant", notificationDelay],
-    ["train", "commuting", notificationDelay],
-    ["exercising", "exercising", notificationDelay]
-  ];
-
-  _.forEach(places, (place) => {
-
-    let [detectorName, situationDescription, delay] = place;
-
-    let need = {
-      needName: `half half: ${situationDescription}`,
-      situation: {
-        detector: DETECTORS[detectorName]._id,
-        number: numberInSituation
-      },
-      toPass: {
-        instruction: `Having a good time ${situationDescription}? Try taking one side of a photo.`
-      },
-      numberNeeded: 2,
-      notificationDelay: delay
-    };
-
-    let callback = {
-      trigger: `cb.numberOfSubmissions("${need.needName}") % 2`,
-      function: completedCallback.toString(),
-    };
-    experience.contributionTypes.push(need);
-    experience.callbacks.push(callback)
-  });
-
-  return experience;
-};
-
-/**
- *
- * @param numberInSituation [Number] Controls asynchronous vs synchronous. Defaults to Asynchronous.
- * @return {*[]}
- */
-const sameSituationContributionTypes = function(
-  {
-    numberInSituation = 1
-  } = {}
-) {
-  return [{
-    needName: 'Warm, Sunny Weather',
-    situation: {
-      detector: DETECTORS.sunny._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you enjoying <span style="color: #0351ff">good weather today?</span> Share a photo of how you are experiencing the sun.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 1,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Shopping for groceries',
-    notificationSubject: 'Inside a grocery store?',
-    notificationText: 'Participate in an experience where you and others are "Grocery Shopping"',
-    situation: {
-      detector: DETECTORS.grocery._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you <span style="color: #0351ff">shopping for groceries?</span> Share a photo of what you are buying or looking at.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Visiting a Cafe',
-    situation: {
-      detector: DETECTORS.coffee._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you <span style="color: #0351ff">at a cafe?</span> Share a photo of yourself with what you purchased, or what you are doing.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Going out for drinks',
-    situation: {
-      detector: DETECTORS.bar._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you out <span style="color: #0351ff">drinking at the bar?</span> Share a photo of yourself at this bar.',
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Eating Japanese Food',
-    situation: {
-      detector: DETECTORS.eating_japanese._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you eating <span style="color: #0351ff">Japanese food?</span> Share a photo of yourself dining at this restaurant.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Religious Worship',
-    situation: {
-      detector: DETECTORS.castle._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you at a <span style="color: #0351ff">center for religious worship?</span> Share a photo of something around you.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Enjoy sunset',
-    situation: {
-      detector: DETECTORS.sunset._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you out during <span style="color: #0351ff">sunset?</span> Share a photo of what the sky looks like where you are.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 1,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Eating Asian Food',
-    situation: {
-      detector: DETECTORS.eating_with_chopsticks._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you <span style="color: #0351ff">eating at an asian restaurant?</span> Share a photo of yourself dining out right now.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Studying at the library',
-    situation: {
-      detector: DETECTORS.library._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Are you spending part of the day <span style="color: #0351ff">reading?</span> Share a photo of what you are doing.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Greenery',
-    situation: {
-      detector: DETECTORS.forest._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: '<span style="color: #0351ff">Are you spending time at a park?</span> Share a photo of what is going on around you.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: 'Rainy Day',
-    situation: {
-      detector: DETECTORS.rainy._id,
-      number: numberInSituation
-    },
-    toPass: {
-      instruction: 'Is it <span style="color: #0351ff">raining</span> today? Share a photo of what it looks like outside.'
-    },
-    numberNeeded: 50,
-    notificationDelay: 1,
-    allowRepeatContributions: true,
-  }, {
-    needName: "Eating some 'Za",
-    situation: {
-      detector: DETECTORS.eating_pizza._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Are you <span style="color: #0351ff">eating pizza</span> today? Share a photo of yourself at the pizza restaurant.',
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: "Eating out",
-    situation: {
-      detector: DETECTORS.restaurant._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Are you <span style="color: #0351ff">eating out</span> today? Share a photo of yourself at the restaurant.',
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }, {
-    needName: "Eating Big Bites",
-    situation: {
-      detector: DETECTORS.big_bite_restaurant._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Are you <span style="color: #0351ff">eating burritos, sandwiches, or burgers</span> today? Share a photo of yourself at the restaurant.',
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-    allowRepeatContributions: true,
-  }];
-};
-
-const halfhalfEmbodiedContributionTypes = function() {
-  return [{
-    needName: 'Hand Silhouette',
-    situation: {
-      detector: DETECTORS.sunny._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Take a photo, holding your hand towards the sky, covering the sun.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-hands-in-front.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 1,
-  }, {
-    needName: 'I eat with my hands',
-    situation: {
-      detector: DETECTORS.grocery._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Take a photo, holding a fruit or vegetable outstretched with your hands.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-fruit-in-hand.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-  }, {
-    needName: 'Coffee Date',
-    situation: {
-      detector: DETECTORS.coffee._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Are you at a cafe? Take a photo, holding your cup, mug, or plate towards the center of the screen.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-cafe.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-  }, {
-    needName: 'Raise a glass',
-    situation: {
-      detector: DETECTORS.bar._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'What are you drinking? Take a photo, while raising your glass or bottle in front of you.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-cheers.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-  }, {
-    needName: 'Itadakimasu (I humbly receive this meal)',
-    situation: {
-      detector: DETECTORS.eating_japanese._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Take a photo, while holding chopsticks in your hand, saying "Itadakimasu" which translates to "I humbly receive this meal"',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-itadakimasu.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-  }, {
-    needName: 'Religious Architecture',
-    situation: {
-      detector: DETECTORS.castle._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Do you notice the details of religious buildings? Do so now, by outstretching your hand and pointing out of the elements that stick out to you most.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-religious-building.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: 'Touch a sunset',
-    situation: {
-      detector: DETECTORS.sunset._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'What does the sunset look like where you are? Find a good view of the sunset. Then, take a photo, with your hands outstretched towards the sun.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-sunset-heart.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 1,
-  }, {
-    needName: 'Eating with Chopsticks',
-    situation: {
-      detector: DETECTORS.eating_with_chopsticks._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'What can you pick up using chopsticks? Take a photo of what you are eating, holding your chopsticks.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-holding-chopsticks.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90,
-  }, {
-    needName: 'reading a book',
-    situation: {
-      detector: DETECTORS.library._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Sorry to interrupt your reading! Find the nearest book, and take a photo holding up the book to your face.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-book-face.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: 'Hold a plant',
-    situation: {
-      detector: DETECTORS.forest._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Find a plant in the <span style="color: #0351ff">park or garden</span>. Take a photo, with your hand shaped as a half-circle.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-hand-circles-flower.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: 'Feet towards the trees',
-    situation: {
-      detector: DETECTORS.forest._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Find a patch of grass to lay your back on. Then, raise your feet. Take a photo of your foot stretching high into the sky',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-feet-towards-trees.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: 'Leaf Mask',
-    situation: {
-      detector: DETECTORS.forest._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Find a leaf in the park. Take a photo of the leaf covering your face, like it was a mask.',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-leaf-face.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: 'Puddles',
-    situation: {
-      detector: DETECTORS.rainy._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: 'Is it <span style="color: #0351ff">raining</span> today? Find a <span style="color: #0351ff">puddle</span> on the ground. Take a photo of yourself, stomping on the puddle!',
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-puddle-feet.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 1,
-  }, {
-    needName: "Slice of 'Za",
-    situation: {
-      detector: DETECTORS.eating_pizza._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: `Did you order <span style="color: #0351ff">pizza</span>? Hold up a <span style="color: #0351ff">slice of 'Za</span> and take a photo of half the slice!`,
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-pizza-slice.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: "Do you take cream with that",
-    situation: {
-      detector: DETECTORS.coffee._id, // any place that has cups (cafes + bars + restaurants)
-      number: '1'
-    },
-    toPass: {
-      instruction: `Do you have <span style="color: #0351ff">a cup or glass</span> you are drinking? Take a photo with it in the middle of the picture. You can even try to <span style="color: #0351ff">pour some extra "cream"</span> into it too!`,
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-teasing-lotion-in-a-cup.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: "Eat from the same bowl", // bowl? Plate?  (basically all restaurants)
-    situation: {
-      detector: DETECTORS.restaurant._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: `Are you <span style="color: #0351ff">eating out</span> right now? Take a photo of yourself holding up <span style="color: #0351ff">a bowl or plate</span> to the middle of the screen.`,
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-bowls-up.jpg'
-    },
-    numberNeeded: 50,
-    notificationDelay: 90
-  }, {
-    needName: "Big Bites", // Any restaurant that would serve something you'd eat with your hands (burrito, tacos, hotdogs, sandwiches, wraps, burgers, tradamerican, newamerican )
-    situation: {
-      detector: DETECTORS.big_bite_restaurant._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: `Are you <span style="color: #0351ff">eating food that would require a big bite</span> right now? Take a photo of yourself <span style="color: #0351ff">holding up your food</span> to the middle of the screen.`,
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-big-bite.jpg'
-    },
-    numberNeeded: 50,
-  }, {
-    needName: "Friday Night",
-    situation: {
-      detector: DETECTORS.big_bite_restaurant._id,
-      number: '1'
-    },
-    toPass: {
-      instruction: `Are you <span style="color: #0351ff">eating out</span> right now? Take a photo of yourself holding up <span style="color: #0351ff">a bowl or plate</span> to the middle of the screen.`,
-      exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-bowls-up.jpg'
-    },
-    numberNeeded: 50,
-  }]
-};
-
-const create24hoursContributionTypes = function(toPassConstructor, numberNeeded) {
-  let needs = [];
-  for (i = 0; i < 24; i++) {
-    let need = {
-      needName: `hour ${i}`,
-      situation: {
-        detector: DETECTORS[`hour${i}`]._id,
-        number: 1
-      },
-      toPass: toPassConstructor(i),
-      numberNeeded: numberNeeded,
-      notificationDelay: 1
-    };
-    needs.push(need);
-  }
-  return needs;
-};
-
-/**
- *
- * @param contributionTypes
- * @param triggerTemplate [String] should be written as a string, with ES6 templating syntax
- *        i.e. "cb.newSubmission(\"${need.needName}\")"
- *        If using templating syntax, you have access to the each individual need object
- * @param sendNotificationFn
- */
-const notifCbForMultiNeeds = function(contributionTypes, triggerTemplate, sendNotificationFn) {
-  return contributionTypes.map((need) => {
-    return {
-      trigger: eval('`' + triggerTemplate + '`'),
-      function: sendNotificationFn.toString()
-    };
-  });
-};
-
-/** halfhalfRespawnAndNotify:
- * This is a helper function that generates a callback function definition
- * The callback will respawn or create a duplicate of the need that just completed,
- * while also sending notifications to the participants of that need.
- *
- * This function makes strong assumptions about how your OCE contributionTypes are written.
- * i.e. need.needName = 'Name of my need 1'
- * i.e. need.needName = 'Hand Silhouette 1'
- *
- * @param subject [String] subject of notification
- * @param text [String] accompanying subtext of notification
- * @return {any} A function
- */
-const halfhalfRespawnAndNotify = function(subject, text) {
-  const functionTemplate = function (sub) {
-    let contributionTypes = Incidents.findOne(sub.iid).contributionTypes;
-    let need = contributionTypes.find((x) => {
-      return x.needName === sub.needName;
-    });
-
-    // Convert Need Name i to Need Name i+1
-    let splitName = sub.needName.split(' ');
-    let iPlus1 = Number(splitName.pop()) + 1;
-    splitName.push(iPlus1);
-    let newNeedName = splitName.join(' ');
-
-    need.needName = newNeedName;
-    addContribution(sub.iid, need);
-
-    let participants = Submissions.find({
-      iid: sub.iid,
-      needName: sub.needName
-    }).map((submission) => {
-      return submission.uid;
-    });
-
-    notify(participants, sub.iid, '${subject}', '${text}', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-  };
-  return eval('`'+functionTemplate.toString()+'`');
-};
-
-const sendNotificationNew24HourPhotoAlbumSub = function(sub) {
-  let uids = Submissions.find({ iid: sub.iid }).fetch().map(function (x) {
-    return x.uid;
-  });
-
-  notify(uids, sub.iid, 'Someone added to the 24 hour photo album. Click here to see progress on the album.', '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-};
-
-let sendNotificationScavenger = function (sub) {
-  let uids = Submissions.find({ iid: sub.iid }).fetch().map(function (x) {
-    return x.uid;
-  });
-
-  notify(uids, sub.iid, 'Wooh! All the scavenger hunt items were found. Click here to see all of them.', '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-};
-
-let sendNotificationSunset = function (sub) {
-  let uids = Submissions.find({ iid: sub.iid }).fetch().map(function (x) {
-    return x.uid;
-  });
-
-  notify(uids, sub.iid, 'Our sunset timelapse is complete! Click here to see it.', '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-};
-
-let sendNotificationFoodFight = function (sub) {
-  let uids = Submissions.find({ iid: sub.iid }).fetch().map(function (x) {
-    return x.uid;
-  });
-  notify(uids, sub.iid, 'Wooh! Both participants have attacked each other with food pics', '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-};
-
-const sendNotificationTwoHalvesCompleted = function(sub) {
-  console.log("Another pair of halves completed a photo");
-
-  let submissions = Submissions.find({
-    iid: sub.iid,
-    needName: sub.needName
-  }).fetch();
-
-  let participants = submissions.map((submission) => { return submission.uid; });
-
-  notify(participants, sub.iid,
-    `Two people completed a half half photo`,
-    `See the results under ${sub.needName}`,
-    '/apicustomresults/' + sub.iid + '/' + sub.eid);
-};
-
-let EXPERIENCES = {
-  // bumped: createBumped(),
-  storyTime: createStorytime(0),
-  storyTime1: createStorytime(1),
-  storyTime2: createStorytime(2),
-  // storyTime3: createStorytime(3),
-  // storyTime4: createStorytime(4),
-  // storyTime5: createStorytime(5),
-  // storyTime6: createStorytime(6),
-  // storyTime7: createStorytime(7),
-  // independentStorybook: createIndependentStorybook(),
-  // sunset: {
-  //   _id: Random.id(),
-  //   name: 'Sunset',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'sunset',
-  //   contributionTypes: [{
-  //     needName: 'sunset',
-  //     situation: {
-  //       detector: DETECTORS.sunset._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Take a photo of the sunset!'
-  //     },
-  //     numberNeeded: 20,
-  //     notificationDelay: 1,
-  //   }],
-  //   description: 'Create a timelapse of the sunset with others around the country',
-  //   notificationText: 'Take a photo of the sunset!',
-  //   callbacks: [{
-  //     trigger: 'cb.incidentFinished()',
-  //     function: sendNotificationSunset.toString()
-  //   }]
-  // },
-  // halfhalf24: {
-  //   _id: Random.id(),
-  //   name: 'Half Half over 24 hours',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults', // FIXME(rlouie): should be a template grouped by time
-  //   contributionTypes: create24hoursContributionTypes(
-  //     function(i) {
-  //       let zpad_i = ("00" + i).slice(-2);
-  //       let toPass = {
-  //         instruction: `<span style="color: #0351ff">This experience is for testing the Half Half Photo Experience!</span><b> Take a picture of what you are doing today at hour ${zpad_i}:00 today.</b>`
-  //       };
-  //       return toPass;
-  //     },
-  //     10
-  //   ),
-  //   description: 'Create a photo collage of what you and others are doing at each of the hours in a day',
-  //   notificationText: 'Take a photo of what you are doing at this hour',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: sendNotificationNew24HourPhotoAlbumSub.toString()
-  //   }]
-  // },
-  // halfhalfAsynch: createHalfHalf(),
-  // halfhalfSynch: createHalfHalf({numberInSituation: 2}),
-  // halfhalfDay: {
-  //   _id: Random.id(),
-  //   name: 'Half Half Daytime',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: [{
-  //     needName: 'half half: daytime', // FIXME: make more semantically meaningful
-  //     situation: {
-  //       detector: DETECTORS.daytime._id,  // For testing during workday
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Take a photo of like Half Half Travel!'
-  //     },
-  //     numberNeeded: 50, // arbitrarily high for a study
-  //     notificationDelay: 1,
-  //   }],
-  //   description: 'Create adventures that meet halfway! Ready to live in a parallel with someone else?',
-  //   notificationText: 'Participate in Half Half Travel!',
-  //   callbacks: [{
-  //     trigger: 'cb.numberOfSubmissions("half half: daytime") % 2 === 0',
-  //     function: sendNotificationTwoHalvesCompleted.toString()
-  //   }]
-  // },
-  // halfhalfNight: {
-  //   _id: Random.id(),
-  //   name: 'Half Half Nighttime',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: [{
-  //     needName: 'half half: nighttime', // FIXME: make more semantically meaningful
-  //     situation: {
-  //       detector: DETECTORS.night._id,  // For testing during evening
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Take a photo of like Half Half Travel!'
-  //     },
-  //     numberNeeded: 50, // arbitrarily high for a study
-  //     notificationDelay: 1, // no need to delay if its daytime outside
-  //   }],
-  //   description: 'Create adventures that meet halfway! Ready to live in a parallel with someone else?',
-  //   notificationText: 'Participate in Half Half Travel!',
-  //   callbacks: [{
-  //     trigger: 'cb.numberOfSubmissions("half half: nighttime") % 2 === 0',
-  //     function: sendNotificationTwoHalvesCompleted.toString()
-  //   }]
-  // },
-  halfhalf_sunny: {
-    _id: Random.id(),
-    name: 'Hand Silhouette',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Hand Silhouette 1',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.sunny),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Is the <span style="color: #0351ff">weather clear and sunny</span> where you are? Take a photo, <span style="color: #0351ff">holding your hand towards the sky, covering the sun.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-hands-in-front.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'Use the sun to make a silhouette of your hand',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A hand silhouette was completed','View the photo').toString()
-    }]
-  },
-  halfhalf_grocery: {
-    _id: Random.id(),
-    name: 'Grocery Buddies',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Grocery Buddies 1',
-      notificationSubject: 'Inside a grocery store?',
-      notificationText: 'Share an experience with others who are also grocery shopping',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.grocery),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you at the <span style="color: #0351ff">grocery store</span>? Take a photo, <span style="color: #0351ff">holding a fruit or vegetable</span> outstretched with your hands.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-fruit-in-hand.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While shopping for groceries, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Grocery Buddies photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_coffee: {
-    _id: Random.id(),
-    name: 'Coffee Date',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Coffee Date 1',
-      notificationSubject: 'Inside a coffee shop?',
-      notificationText: 'Share an experience with others who are also at a coffee shop',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.coffee),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you at a <span style="color: #0351ff">cafe</span>? Take a photo, <span style="color: #0351ff">holding your cup, mug, or plate</span> towards the center of the screen.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-cafe.jpg'
-      },
-      numberNeeded: 2,
-      notificationDelay: 90,
-    }]),
-    description: 'While enjoying a cafe beverage, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Coffee Date photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_bar: {
-    _id: Random.id(),
-    name: 'Cheers',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Cheers 1',
-      notificationSubject: 'Drinking at a bar?',
-      notificationText: 'Share an experience with others who are also drinking at a bar',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.bar),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'What are you <span style="color: #0351ff">drinking at the bar</span>? Take a photo, while <span style="color: #0351ff">raising your glass or bottle</span> in front of you.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-cheers.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While enjoying your drink, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Cheers photo completed','View the photo').toString()
-    }]
-  },
-  // halfhalf_japanese: {
-  //   _id: Random.id(),
-  //   name: 'Itadakimasu',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-  //     // needName MUST have structure "My Need Name XYZ"
-  //     needName: 'Itadakimasu 1',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.eating_japanese),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Take a photo, while <span style="color: #0351ff">holding chopsticks in your hand</span>, saying "Itadakimasu" which translates to "I humbly receive this meal"',
-  //       exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-itadakimasu.jpg'
-  //     },
-  //     numberNeeded: 2,
-  //     notificationDelay: 90,
-  //   }]),
-  //   description: 'While eating Japanese Food, create a half half photo.',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: '(cb.numberOfSubmissions() % 2) === 0',
-  //     function: halfhalfRespawnAndNotify('A Itadakimasu photo completed','View the photo').toString()
-  //   }]
-  // },
-  halfhalf_religious: {
-    _id: Random.id(),
-    name: 'Religious Architecture',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Religious Architecture 1',
-      notificationSubject: 'Visiting a place of worship?',
-      notificationText: 'Share an experience with others who are also visiting a place of worship',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.castle),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Do you notice the <span style="color: #0351ff">details of the religious building</span> near you? Do so now, by outstretching your hand and pointing out of the elements that stick out to you most.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-religious-building.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While visiting a place of worship, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Religious Architecture photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_sunset: {
-    _id: Random.id(),
-    name: 'Sunset Together',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Sunset Together 1',
-      notificationSubject: 'Can you see the sunset?',
-      notificationText: 'Share an experience with others who are also watching the sunset',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.sunset),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'What does the <span style="color: #0351ff">sunset</span> look like where you are? Find a good view; then, take a photo, with your hands outstretched towards the sun.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-sunset-heart.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'While looking up at the sunset, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Sunset Together photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_asian: {
-    _id: Random.id(),
-    name: 'Eating with Chopsticks',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Eating with Chopsticks 1',
-      notificationSubject: 'Eating at an asian restaurant?',
-      notificationText: 'Share an experience with others who are also eating asian food',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.eating_with_chopsticks),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you eating <span style="color: #0351ff">asian food</span> right now? Take a photo of what you are eating, <span style="color: #0351ff">holding your chopsticks.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-holding-chopsticks.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While eating asian food, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('An Eating with Chopsticks photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_books: {
-    _id: Random.id(),
-    name: 'Book Buddies',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Book Buddies 1',
-      notificationSubject: 'Are you at a library?',
-      notificationText: 'Share an experience with others who are also at the library',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.library),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Sorry to interrupt your <span style="color: #0351ff">reading</span>! Find the nearest book, and take a photo <span style="color: #0351ff">holding up the book to your face.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-book-face.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While reading a book, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Book Buddies photo completed','View the photo').toString()
-    }]
-  },
-  // halfhalf_plantcircle: {
-  //   _id: Random.id(),
-  //   name: 'Hold a plant',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-  //     // needName MUST have structure "My Need Name XYZ"
-  //     needName: 'Hold a plant 1',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.forest),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Find a plant in the <span style="color: #0351ff">park or garden</span>. Take a photo, with your hand shaped as a half-circle.',
-  //       exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-hand-circles-flower.jpg'
-  //     },
-  //     numberNeeded: 2,
-  //     notificationDelay: 90,
-  //   }]),
-  //   description: 'While in the park, create a half half photo.',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: '(cb.numberOfSubmissions() % 2) === 0',
-  //     function: halfhalfRespawnAndNotify('A Plant Circle photo completed','View the photo').toString()
-  //   }]
-  // },
-  // halfhalf_feettotrees: {
-  //   _id: Random.id(),
-  //   name: 'Feet to the trees',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-  //     // needName MUST have structure "My Need Name XYZ"
-  //     needName: 'Feet to the trees 1',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.forest),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Find a <span style="color: #0351ff">patch of grass to lay your back on</span>. Then, raise your feet. Take a photo of your foot stretching high into the sky',
-  //       exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-feet-towards-trees.jpg'
-  //     },
-  //     numberNeeded: 2,
-  //     notificationDelay: 5,
-  //   }]),
-  //   description: 'While in the park, create a half half photo.',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: '(cb.numberOfSubmissions() % 2) === 0',
-  //     function: halfhalfRespawnAndNotify('A Feet to the trees photo completed','View the photo').toString()
-  //   }]
-  // },
-  halfhalf_leakmask: {
-    _id: Random.id(),
-    name: 'Leaf Mask',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Leaf Mask 1',
-      notificationSubject: 'Are you at a park?',
-      notificationText: 'Share an experience with others who are also at a park',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.forest),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Find a <span style="color: #0351ff">leaf in the park</span>. Take a photo of the <span style="color: #0351ff">leaf covering your face, like it was a mask.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-leaf-face.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While in the park, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Feet to the trees photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_puddles: {
-    _id: Random.id(),
-    name: 'Puddle Feet',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Puddle Feet 1',
-      notificationSubject: 'Are you outside while its raining?',
-      notificationText: 'Share an experience with others who are enjoying or enduring the rain',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.rainy),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Is it <span style="color: #0351ff">raining</span> today? Find a <span style="color: #0351ff">puddle</span> on the ground. Take a photo of yourself, stomping on the puddle!',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-puddle-feet.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'With the puddles on a rainy day, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A "Puddle Feet" photo completed','View the photo').toString()
-    }]
-  },
-  // halfhalf_pizza: {
-  //   _id: Random.id(),
-  //   name: "Slice of 'Za",
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-  //     // needName MUST have structure "My Need Name XYZ"
-  //     needName: "Slice of 'Za 1",
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.eating_pizza),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: `Did you order <span style="color: #0351ff">pizza</span>? Hold up a <span style="color: #0351ff">slice of 'Za</span> and take a photo of half the slice!`,
-  //       exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-pizza-slice.jpg'
-  //     },
-  //     numberNeeded: 2,
-  //     notificationDelay: 90,
-  //   }]),
-  //   description: 'While eating pizza, create a half half photo.',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: '(cb.numberOfSubmissions() % 2) === 0',
-  //     function: halfhalfRespawnAndNotify("A \"Slice of 'Za\" photo completed",'View the photo').toString()
-  //   }]
-  // },
-  // halfhalf_creamwiththat: {
-  //   _id: Random.id(),
-  //   name: "Want cream with that",
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-  //     // needName MUST have structure "My Need Name XYZ"
-  //     needName: "Want cream with that 1",
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.coffee), // any place that has cups (cafes + bars + restaurants)
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: `Do you have <span style="color: #0351ff">a cup or glass</span> you are drinking? Take a photo with it in the middle of the picture. You can even try to <span style="color: #0351ff">pour some extra "cream"</span> into it too!`,
-  //       exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-teasing-lotion-in-a-cup.jpg'
-  //     },
-  //     numberNeeded: 2,
-  //     notificationDelay: 60,
-  //   }]),
-  //   description: 'While drinking coffee at a cafe, create a half half photo.',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: '(cb.numberOfSubmissions() % 2) === 0',
-  //     function: halfhalfRespawnAndNotify("A 'Want cream with that' photo completed",'View the photo').toString()
-  //   }]
-  // },
-  // halfhalf_eatfromsamebowl: {
-  //   _id: Random.id(),
-  //   name: "Share a plate",
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-  //     // needName MUST have structure "My Need Name XYZ"
-  //     needName: "Share a plate 1", // bowl? Plate?  (basically all restaurants)
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.restaurant),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: `Are you <span style="color: #0351ff">eating out</span> right now? Take a photo of yourself holding up <span style="color: #0351ff">a bowl or plate</span> to the middle of the screen.`,
-  //       exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-bowls-up.jpg'
-  //     },
-  //     numberNeeded: 2,
-  //     notificationDelay: 60,
-  //   }]),
-  //   description: 'While eating out at a restaurant, create a half half photo.',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: '(cb.numberOfSubmissions() % 2) === 0',
-  //     function: halfhalfRespawnAndNotify("A 'Share a Plate' photo completed",'View the photo').toString()
-  //   }]
-  // },
-  halfhalf_bigbites: {
-    _id: Random.id(),
-    name: "Big Bites",
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('mechanismRich', [{
-      needName: "Big Bites 1", // Any restaurant that would serve something you'd eat with your hands (burrito, tacos, hotdogs, sandwiches, wraps, burgers, tradamerican, newamerican )
-      notificationSubject: 'Eating at a restaurant?',
-      notificationText: 'Share an experience with others who are enjoying big bites of their meal',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.big_bite_restaurant),
-        number: '1'
-      },
-      toPass: {
-        instruction: `Are you <span style="color: #0351ff">eating food that would require a big bite</span> right now? Take a photo of yourself <span style="color: #0351ff">holding up your food</span> to the middle of the screen.`,
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-big-bite.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90, // https://www.quora.com/Whats-the-average-time-that-customers-wait-between-entering-a-restaurant-and-getting-served
-    }]),
-    description: 'While eating some non-trivially sized food, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify("A 'Big Bites' photo completed",'View the photo').toString()
-    }]
-  },
-  // halfhalfEmbodiedMimicry: {
-  //   _id: Random.id(),
-  //   name: 'Body Mirror',
-  //   participateTemplate: 'halfhalfParticipate',
-  //   resultsTemplate: 'halfhalfResults',
-  //   contributionTypes: halfhalfEmbodiedContributionTypes(),
-  //   description: 'With your environment as the shared canvas, pose your body to be the mirror image of a friend',
-  //   notificationText: 'Your situation made you available to participate in Body Mirror!',
-  //   callbacks: notifCbForMultiNeeds(
-  //     halfhalfEmbodiedContributionTypes(),
-  //     "cb.numberOfSubmissions(\"${need.needName}\") % 2 === 0",
-  //     sendNotificationTwoHalvesCompleted)
-  // },
-  // situationaware_sunny: {
-  //   _id: Random.id(),
-  //   name: 'Sunny Days',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Sunny Days',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.sunny),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you enjoying <span style="color: #0351ff">good weather today?</span> Share a photo of how you are experiencing the sun.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 1,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Sunny Days', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_grocery: {
-  //   _id: Random.id(),
-  //   name: 'Feed yourself',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Feed yourself',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.grocery),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">shopping for groceries?</span> Share a photo of what you are buying or looking at.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 90,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Feed yourself', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_cafe: {
-  //   _id: Random.id(),
-  //   name: 'Cafe Days',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Cafe Days',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.coffee),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">at a cafe?</span> Share a photo of yourself with what you purchased, or what you are doing.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 90,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Cafe Days', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_bar: {
-  //   _id: Random.id(),
-  //   name: 'Hit the Bars',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Hit the Bars',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.bar),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you out <span style="color: #0351ff">drinking at the bar?</span> Share a photo of yourself at this bar.',
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 90,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Hit the Bars', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_japanese: {
-  //   _id: Random.id(),
-  //   name: 'Eating Japanese Food',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Eating Japanese Food',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.eating_japanese),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you eating <span style="color: #0351ff">Japanese food?</span> Share a photo of yourself dining at this restaurant.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 90,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Eating Japanese Food', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_religious: {
-  //   _id: Random.id(),
-  //   name: 'Religious Worship',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Religious Worship',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.castle),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you at a <span style="color: #0351ff">center for religious worship?</span> Share a photo of something around you.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 30,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Religious Worship', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_sunset: {
-  //   _id: Random.id(),
-  //   name: 'Catch the sunset',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Catch the sunset',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.sunset),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you out during <span style="color: #0351ff">sunset?</span> Share a photo of what the sky looks like where you are.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 1,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Catch the sunset', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_asian: {
-  //   _id: Random.id(),
-  //   name: 'Eating Asian Food',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Eating Asian Food',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.eating_with_chopsticks),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">eating at an asian restaurant?</span> Share a photo of yourself dining out right now.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 90,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Eating Asian Food', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_books: {
-  //   _id: Random.id(),
-  //   name: 'Reading a book',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Reading a book',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.library),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you spending part of the day <span style="color: #0351ff">reading?</span> Share a photo of what you are doing.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 90,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Reading a book', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_parks: {
-  //   _id: Random.id(),
-  //   name: 'I love parks',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'I love parks',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.forest),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">at a park?</span> Share a photo of what is going on around you.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 15,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for I love parks', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_rainy: {
-  //   _id: Random.id(),
-  //   name: 'Rainy Day',
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: 'Rainy Day',
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.rainy),
-  //       number: 1
-  //     },
-  //     toPass: {
-  //       instruction: 'Is it <span style="color: #0351ff">raining</span> today? Share a photo of what it looks like outside.'
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 1,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Rainy Day', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_pizza: {
-  //   _id: Random.id(),
-  //   name: "Eating some 'Za",
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: "Eating some 'Za",
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.eating_pizza),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">eating pizza</span> today? Share a photo of yourself at the pizza restaurant.',
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 60,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Eating some \'Za', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_eatout: {
-  //   _id: Random.id(),
-  //   name: "Eating out",
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: "Eating out",
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.restaurant),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">eating out</span> today? Share a photo of yourself at the restaurant.',
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 60,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Eating out', 'View the photo').toString()
-  //   }]
-  // },
-  // situationaware_bigbite: {
-  //   _id: Random.id(),
-  //   name: "Eating Big Bites",
-  //   participateTemplate: 'uploadPhoto',
-  //   resultsTemplate: 'photosByCategories',
-  //   contributionTypes: addStaticAffordanceToNeeds('mechanismPoor', [{
-  //     needName: "Eating Big Bites",
-  //     situation: {
-  //       detector: getDetectorUniqueKey(DETECTORS.big_bite_restaurant),
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Are you <span style="color: #0351ff">eating burritos, sandwiches, or burgers</span> today? Share a photo of yourself at the restaurant.',
-  //     },
-  //     numberNeeded: 50,
-  //     notificationDelay: 60,
-  //     allowRepeatContributions: true,
-  //   }]),
-  //   description: 'Appreciate the small moments with others who are doing the same',
-  //   notificationText: 'View this and other available experiences',
-  //   callbacks: [{
-  //     trigger: 'cb.newSubmission()',
-  //     function: notifyUsersInNeed('New moment for Eating Big Bites', 'View the photo').toString()
-  //   }]
-  // },
-  sameSituationAwareness: {
-    _id: Random.id(),
-    name: 'Our Small Moments',
-    participateTemplate: 'uploadPhoto',
-    resultsTemplate: 'photosByCategories',
-    contributionTypes: sameSituationContributionTypes(),
-    description: 'During the small moments, we might be experiencing more together than we let on.',
-    notificationText: 'Your situation made you available to participate in Our Small Moments!',
-    callbacks: notifCbForMultiNeeds(
-      sameSituationContributionTypes(),
-      "cb.newSubmission(\"${need.needName}\")",
-      notifyUsersInNeed('Someone added to Our Small Moments', 'See the results under ${sub.needName}'))
-  },
-  scavengerHunt: {
-    _id: Random.id(),
-    name: 'St. Patrick\'s Day Scavenger Hunt',
-    participateTemplate: 'scavengerHuntParticipate',
-    resultsTemplate: 'scavengerHunt',
-    contributionTypes: [{
-      needName: 'beer',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.beer),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of beer?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 30, // 30 seconds for debugging
-    }, {
-      needName: 'greenProduce',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.produce),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of green vegetables? #leprechaunfood'
-      },
-      numberAllowedToParticipateAtSameTime: 1, // for testing
-      numberNeeded: 5, // for testing
-      notificationDelay: 20, // 20 seconds for debugging
-    }, {
-      needName: 'coins',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.drugstore),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of chocolate gold coins on display?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 10, // 10 seconds for debugging
-    }, {
-      needName: 'leprechaun',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.costume_store),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of a Leprechaun costume?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 15, // 15 seconds for debugging
-    }, {
-      needName: 'irishSign',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.irish),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of an Irish sign?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 1, // 1 seconds for debugging (passing by)
-    }, {
-      needName: 'trimmings',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.hair_salon),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of some Leprechaun beard trimmings?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 10, // 10 seconds for debugging
-    }, {
-      needName: 'liquidGold',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.gas_station),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of liquid gold that Leprechauns use to power their vehicles?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 10, // 10 seconds for debugging
-    }, {
-      needName: 'potOfGold',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.bank),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of a bank where Leprechauns hide their pots of gold?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 10, // 10 seconds for debugging
-    }, {
-      needName: 'rainbow',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.rainbow),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Can you take a photo of a rainbow flag?'
-      },
-      numberNeeded: 1,
-      notificationDelay: 10, // 10 seconds for debugging
-    }],
-    description: 'Find an item for a scavenger hunt',
-    notificationText: 'Help us complete a St. Patrick\'s day scavenger hunt',
-    callbacks: [{
-      trigger: 'cb.incidentFinished()',
-      function: sendNotificationScavenger.toString()
-    }]
-  },
-  // natureHunt: {
-  //   _id: Random.id(),
-  //   name: 'Nature Scavenger Hunt',
-  //   participateTemplate: 'scavengerHuntParticipate',
-  //   resultsTemplate: 'scavengerHunt',
-  //   contributionTypes: [{
-  //     needName: 'tree',
-  //     situation: {
-  //       detector: DETECTORS.forest._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of a tree?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 10, // 10 seconds for debugging
-  //   }, {
-  //     needName: 'leaf',
-  //     situation: {
-  //       detector: DETECTORS.forest._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of a leaf?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 10, // 10 seconds for debugging
-  //   }, {
-  //     needName: 'grass',
-  //     situation: {
-  //       detector: DETECTORS.field._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the grass?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 10, // 10 seconds for debugging
-  //   }, {
-  //     needName: 'lake',
-  //     situation: {
-  //       detector: DETECTORS.lake._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the lake?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 10, // 10 seconds for debugging
-  //   }, {
-  //     needName: 'moon',
-  //     situation: {
-  //       detector: DETECTORS.night._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the moon?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 1, // 1 seconds for debugging
-  //   }, {
-  //     needName: 'sun',
-  //     situation: {
-  //       detector: DETECTORS.sunny._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the sun?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 1, // 1 seconds for debugging
-  //   }, {
-  //     needName: 'blueSky',
-  //     situation: {
-  //       detector: DETECTORS.sunny._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the blue sky?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 1, // 1 seconds for debugging
-  //   }, {
-  //     needName: 'clouds',
-  //     situation: {
-  //       detector: DETECTORS.cloudy._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the clouds?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 1, // 1 seconds for debugging
-  //   }, {
-  //     needName: 'puddle',
-  //     situation: {
-  //       detector: DETECTORS.rainy._id,
-  //       number: '1'
-  //     },
-  //     toPass: {
-  //       instruction: 'Can you take a photo of the puddle?'
-  //     },
-  //     numberNeeded: 1,
-  //     notificationDelay: 1, // 1 seconds for debugging
-  //   }],
-  //   description: 'Find an item for a scavenger hunt',
-  //   notificationText: 'Help us out with our nature scavenger hunt',
-  //   callbacks: [{
-  //     trigger: 'cb.incidentFinished()',
-  //     function: sendNotificationScavenger.toString()
-  //   }]
-  // },
-  // foodfight: {
-  //   _id: Random.id(),
-  //   name: "Food Fight!",
-  //   participateTemplate: "scavengerHuntParticipate",
-  //   resultsTemplate: "scavengerHunt",
-  //   contributionTypes: [
-  //     {
-  //       needName: "foodPhoto",
-  //       situation: {
-  //         detector: DETECTORS.restaurant._id,
-  //         number: 1
-  //       },
-  //       toPass: {
-  //         instruction: "Can you take a photo of what you're eating?"
-  //       },
-  //       numberNeeded: 1
-  //     },
-  //     {
-  //       needName: "foodPhoto",
-  //       situation: {
-  //         detector: DETECTORS.restaurant._id,
-  //         number: 1
-  //       },
-  //       toPass: {
-  //         instruction: "Can you take a photo of what you're eating?"
-  //       },
-  //       numberNeeded: 1
-  //     }
-  //   ],
-  //   description: "Food fight!",
-  //   notificationText: "Food fight!",
-  //   callbacks: [{
-  //       trigger: "cb.incidentFinished()",
-  //       function: sendNotificationFoodFight.toString()
-  //   }]
-  // }
-};
-
 import CHI20_DTR_EXPERIENCES from "./summer19study/dtr_experiences";
 import CHI20_Olin_EXPERIENCES from "./summer19study/olin_experiences";
+import SPRING18_EXPERIENCES from "./spring18study/four_interaction_structures";
+import SUMMER18_BETA from "./summer18beta/beta_experiences";
+import SUMMER18_EXPERIENCES from "./summer18study/summer18experiences";
+
+let EXPERIENCES = Object.assign({},
+  CHI20_DTR_EXPERIENCES,
+  CHI20_Olin_EXPERIENCES,
+  SPRING18_EXPERIENCES,
+  SUMMER18_BETA,
+  SUMMER18_EXPERIENCES,
+);
 
 export const CONSTANTS = {
   'LOCATIONS': LOCATIONS,
   'USERS': USERS,
-  // Comment out if you would like to only test specific experiences
-  // 'EXPERIENCES': (({ halfhalfEmbodiedMimicry }) => ({ halfhalfEmbodiedMimicry }))(EXPERIENCES),
-  'EXPERIENCES': Object.assign({}, EXPERIENCES, CHI20_DTR_EXPERIENCES, CHI20_Olin_EXPERIENCES),
-  // 'EXPERIENCES': CHI20_DTR_EXPERIENCES,
+  'EXPERIENCES': EXPERIENCES,
   'DETECTORS': DETECTORS
 };
